Refactor models
- Separate model definitions
