--- src/renderer/store/models.js
@@ -1,331 +1,0 @@
-import { Model } from '@vuex-orm/core';
-import {
-  NOTE_ON, NOTE_OFF, MODULATION,
-  NOTE, SPECTROGRAM, AUDIO_BUFFER, CLIP,
-  TRACK, SONG, APP, PIANO_ROLL, ARRANGEMENT
-} from '../constants/model-types';
-import { makeMandatory, setHotkeysScope } from './utils';
-import { bpm, beatsPerBar, ticksPerBeat } from '../constants/defaults';
-import { SELECT } from '../constants/mouse-modes';
-import { random, sum, flatten } from 'lodash';
-import { SCALE_COLORS } from '../constants/colors';
-
-class BaseModel extends Model {
-  get path () {
-    if (!this.parent) return [this];
-    return [...this.parent.path, this];
-  }
-  get absoluteTime () {
-    const offsetTimes = this.path.map(modelInstance => {
-      const offset = modelInstance.offsetTime;
-      return offset == null ? 0 : offset;
-    });
-    return sum(offsetTimes);
-  }
-}
-
-export class NoteOn extends BaseModel {
-  static get entity () {
-    return 'noteOns';
-  }
-
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      noteId: this.attr(null),
-      type: this.string(NOTE_ON),
-      noteOnVelocity: this.number(0.5), // from 0.0 to 1.0.
-      pitchBend: this.number(0), // in midi note number. Negative float is acceptable.
-      timbre: this.number(0.5), // from 0.0 to 1.0.
-      pressure: this.number(0.5), // from 0.0 to 1.0.
-      selected: this.boolean(false)
-    };
-  }
-  get parent () {
-    return Note.query().whereId(this.noteId).first();
-  }
-}
-
-export class Modulation extends BaseModel {
-  static get entity () {
-    return 'modulations';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      noteId: this.attr(null),
-      type: this.string(MODULATION),
-      offsetTime: this.number(1), // in tick
-      pitchBend: this.number(null).nullable(), // in midi note number. Negative float is acceptable.
-      pressure: this.number(null).nullable(), // from 0.0 to 1.0.
-      timbre: this.number(null).nullable(), // from 0.0 to 1.0.
-      selected: this.boolean(false)
-    };
-  }
-  get parent () {
-    return Note.query().whereId(this.noteId).first();
-  }
-}
-
-export class NoteOff extends BaseModel {
-  static get entity () {
-    return 'noteOffs';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      noteId: this.attr(null),
-      type: this.string(NOTE_OFF),
-      offsetTime: this.number(1), // in tick
-      noteOffVelocity: this.number(0),
-      pitchBend: this.number(null).nullable(), // in midi note number. Negative float is acceptable.
-      pressure: this.number(null).nullable(), // from 0.0 to 1.0.
-      timbre: this.number(null).nullable(), // from 0.0 to 1.0.
-      selected: this.boolean(false)
-    };
-  }
-  get parent () {
-    return Note.query().whereId(this.noteId).first();
-  }
-}
-
-export class Note extends BaseModel {
-  static get entity () {
-    return 'notes';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(NOTE),
-      clipId: this.attr(null, makeMandatory('clipId')),
-      noteNumber: this.number(60),
-      offsetTime: this.number(0), // in tick
-      noteOn: this.hasOne(NoteOn, 'noteId'),
-      noteOff: this.hasOne(NoteOff, 'noteId'),
-      modulations: this.hasMany(Modulation, 'noteId'),
-      selected: this.boolean(false),
-      interpolation: this.string('LINEAR')
-    };
-  }
-  get noteActions () {
-    return [
-      this.noteOn,
-      ...this.modulations,
-      this.noteOff
-    ];
-  }
-  get pitchTransition () {
-    const pitchBendMods = this.modulations.filter(mod => mod.pitchBend !== null);
-    const noteActions = [{ offsetTime: 0, ...this.noteOn }, ...pitchBendMods, this.noteOff && this.noteOff].filter(v => v);
-    return noteActions.map(noteAction => ({
-      offsetTime: this.offsetTime + noteAction.offsetTime,
-      pitch: this.noteNumber + noteAction.pitchBend,
-      id: noteAction.id,
-      type: noteAction.type
-    }));
-  }
-  get parent () {
-    return Clip.query().whereId(this.clipId).first();
-  }
-}
-
-export class AudioBuffer extends BaseModel {
-  static get entity () {
-    return 'audioBuffers';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(AUDIO_BUFFER),
-      clipId: this.attr(null, makeMandatory('clipId')),
-      offsetTime: this.number(0),
-      data: this.attr(null),
-      originalFilePath: this.string(''),
-      spectrogram: this.hasOne(Spectrogram, 'audioBufferId')
-    };
-  }
-  get parent () {
-    return Clip.query().whereId(this.clipId).first();
-  }
-}
-
-export class Spectrogram extends BaseModel {
-  static get entity () {
-    return 'spectrograms';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(SPECTROGRAM),
-      audioBufferId: this.attr(null),
-      times: this.attr([]),
-      freqs: this.attr([]),
-      magnitude2d: this.attr([[]])
-    };
-  }
-  get parent () {
-    return AudioBuffer.query().whereId(this.audioBufferId).first();
-  }
-}
-
-export class Clip extends BaseModel {
-  static get entity () {
-    return 'clips';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(CLIP),
-      offsetTime: this.number(0), // in tick
-      duration: this.number(4 * beatsPerBar * ticksPerBeat), // in tick. default is 4 bars
-      notes: this.hasMany(Note, 'clipId'),
-      audioBuffer: this.hasOne(AudioBuffer, 'clipId'),
-      beatsPerBar: this.attr([{ barIdx: 0, val: beatsPerBar }]),
-      selected: this.boolean(false),
-      trackId: this.attr(null),
-      color: this.attr(() => SCALE_COLORS.hHelmholtz[random(11)])
-    };
-  };
-  get parent () {
-    return Track.query().whereId(this.trackId).first();
-  }
-  get startTime () {
-    return this.absoluteTime;
-  }
-  get endTime () {
-    return this.absoluteTime + this.duration;
-  }
-  get selectedNoteIds () {
-    return this.notes.filter(note => note.selected).map(note => note.id);
-  }
-  get someNotesAreSelected () {
-    return this.selectedNoteIds.length !== 0;
-  }
-}
-
-export class Track extends BaseModel {
-  static get entity () {
-    return 'tracks';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(TRACK),
-      speed: this.attr([{ offsetTime: 0, val: 1 }]),
-      selected: this.boolean(false),
-      songId: this.attr(null),
-      clips: this.hasMany(Clip, 'trackId')
-    };
-  }
-  get parent () {
-    return Song.query().whereId(this.songId).first();
-  }
-}
-
-export class Song extends BaseModel {
-  static get entity () {
-    return 'songs';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(SONG),
-      bpm: this.number(bpm),
-      ticksPerBeat: this.number(ticksPerBeat),
-      tracks: this.hasMany(Track, 'songId')
-    };
-  }
-  get parent () {
-    return null;
-  }
-}
-
-export class PianoRoll extends BaseModel {
-  static get entity () {
-    return 'pianoRoll';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(PIANO_ROLL),
-      appId: this.attr(null, makeMandatory('appId')),
-      gridOpacity: this.attr(1),
-      spectrogramOpacity: this.attr(1),
-      mouseMode: this.string(SELECT)
-    };
-  }
-  get parent () {
-    return App.query().whereId(this.appId).first();
-  }
-  get clips () {
-    return Clip.query().where('selected', true).withAllRecursive().get();
-  }
-  get notes () {
-    return flatten(this.clips.map(clip => clip.notes));
-  }
-  get selectedNoteIds () {
-    return flatten(this.clips.map(clip => clip.selectedNoteIds));
-  }
-  get someNotesAreSelected () {
-    return this.clips.some(clip => clip.someNotesAreSelected);
-  }
-  get audioBuffers () {
-    return this.clips.map(clip => clip.audioBuffer).filter(v => v);
-  }
-  get spectrograms () {
-    return this.audioBuffers.map(audioBuffer => audioBuffer.spectrogram).filter(v => v);
-  }
-  get beatsPerBar () {
-    return this.clips[0].beatsPerBar[0].val;
-  }
-  get ticksPerBeat () {
-    return Song.query().last().ticksPerBeat;
-  }
-  get displayRange () {
-    const start = Math.min(this.clips.map(clip => clip.startTime));
-    const end = Math.max(this.clips.map(clip => clip.endTime));
-    return { start, end };
-  }
-  get totalTicks () {
-    return this.displayRange.end - this.displayRange.start;
-  }
-  get totalBeats () {
-    return this.totalTicks / this.ticksPerBeat;
-  }
-  get totalBars () {
-    return this.totalBeats / this.beatsPerBar;
-  }
-}
-
-export class Arrangement extends BaseModel {
-  static get entity () {
-    return 'arrangement';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(ARRANGEMENT),
-      appId: this.attr(null, makeMandatory('appId')),
-      mouseMode: this.string(SELECT)
-    };
-  }
-  get parent () {
-    return App.query().whereId(this.appId).first();
-  }
-}
-
-export class App extends BaseModel {
-  static get entity () {
-    return 'app';
-  }
-  static fields () {
-    return {
-      id: this.attr(null, makeMandatory('id')),
-      type: this.string(APP),
-      selectedZone: this.string(PIANO_ROLL, setHotkeysScope)
-    };
-  }
-  get parent () {
-    return null;
-  }
-}
