+++ src/renderer/store/models/index.js
@@ -1,0 +1,13 @@
+export { default as BaseModel } from './BaseModel';
+export { default as Note } from './Note';
+export { default as NoteOn } from './NoteOn';
+export { default as NoteOff } from './NoteOff';
+export { default as Modulation } from './Modulation';
+export { default as Clip } from './Clip';
+export { default as AudioBuffer } from './AudioBuffer';
+export { default as Spectrogram } from './Spectrogram';
+export { default as Track } from './Track';
+export { default as Song } from './Song';
+export { default as PianoRoll } from './PianoRoll';
+export { default as Arrangement } from './Arrangement';
+export { default as App } from './App';
